Tab bar is working. Modal content is moved and working
