+++ frontend/src/js/components/modals/CachedFiles/Tabs/FileTab.jsx
@@ -1,19 +1,27 @@
-import React, { Component } from 'react'
-import { Button, Row, Col, Glyphicon, FormControl, InputGroup } from 'react-bootstrap';
+import React, { Component } from 'react';
+import { Modal, Button, Row, Col, Glyphicon, FormGroup, FormControl, ControlLabel, InputGroup } from 'react-bootstrap';
 import _ from 'lodash';
+import Dialog from 'react-bootstrap-dialog';
 import { DropTarget, DragSource } from 'react-dnd';
-import DragAndDropTypes from 'constants/DragAndDropTypes';
 import { findDOMNode } from 'react-dom';
-import { cleanPath } from '../CachedFiles.jsx'
-import DimensionSlider from './DimensionSlider/DimensionSlider.jsx';
 
+import DimensionSlider from './../DimensionSlider/DimensionSlider.jsx';
+import FileExplorer from '../../FileExplorer/FileExplorer.jsx';
+import DragAndDropTypes from 'constants/DragAndDropTypes';
+import TabBar from './TabBar.jsx'
+
+import '../CachedFiles.scss';
+
+function cleanPath(path) {
+    return `/${path.split('/').filter(segment => segment).join('/')}`;
+}
 
 const HISTORY_KEY = "variable_history_files"
 const BOOKMARK_KEY = "variable_bookmark_files"
 
-class FileTab extends Component{
-    constructor(props){
-        super(props)
+class FileTab extends Component {
+    constructor(props) {
+        super(props);
         let history_files;
         let bookmark_files;
         try {
@@ -34,8 +42,10 @@ class FileTab extends Component{
         catch(e){
             bookmark_files = []
         }
-        this.state ={
+        
+        this.state = {
             showFileExplorer: false,
+            showRedefineVariableModal: false,
             selectedFile: '',
             historyFiles: history_files,
             bookmarkFiles: bookmark_files,
@@ -45,8 +55,104 @@ class FileTab extends Component{
             selectedVariableName: '',
             redefinedVariableName: '',
             temporaryRedefinedVariableName: '',
-            dimension: null,
+            dimension: null
         }
+    }
+
+    get selectedFilePath() {
+        return !this.state.selectedFile ? '' : cleanPath(this.state.selectedFile.path + '/' + this.state.selectedFile.name);
+    }
+
+    get variableName() {
+        if (this.state.redefinedVariableName) {
+            return this.state.redefinedVariableName;
+        }
+        return !this.state.selectedVariableName ? '' : this.state.selectedVariableName;
+    }
+
+    componentDidUpdate(prevProps, prevState) {
+        if (this.state.selectedVariableName !== prevState.selectedVariableName || (this.state.selectedVariableName && !this.state.selectedVariable)) {
+            let selectedVariable, dimension;
+            if (this.state.variablesAxes[0][this.state.selectedVariableName]) {
+                // it's a variablevar 
+                selectedVariable = this.state.variablesAxes[0][this.state.selectedVariableName];
+                dimension = selectedVariable.axisList.map((axisName) => {
+                    return { axisName };
+                })
+            }
+            else {
+                // it's a axis
+                selectedVariable = this.state.variablesAxes[1][this.state.selectedVariableName];
+                dimension = { axisName: this.state.selectedVariableName };
+            }
+            this.setState({
+                selectedVariable,
+                dimension
+            });
+        }
+    }
+
+    loadVariable() {
+        let variable = this.state.selectedVariableName;
+        let filename = this.state.selectedFile;
+        let path = this.selectedFilePath;
+
+        let var_obj = {};
+        var_obj[this.variableName] = {
+            cdms_var_name: variable,
+            axisList: this.state.selectedVariable.axisList,
+            filename: filename,
+            path: path,
+            dimension: this.state.dimension
+        };
+        this.props.loadVariables([var_obj]);
+        this.setState({ redefinedVariableName: '' });
+    }
+
+    load() {
+        return this.variableNameExists()
+            .then((result) => {
+                if (result) {
+                    return new Promise((resolve, reject) => {
+                        this.refs.dialog.show({
+                            title: 'Variable exists',
+                            body: `The variable name ${this.variableName} already exists, rename or overwrite existing varaible`,
+                            actions: [
+                                Dialog.OKAction(() => resolve())
+                            ]
+                        })
+                    })
+                        .then(() => this.redefineVariableName());
+                }
+            })
+            .then(() => {
+                return this.loadVariable()
+            })
+    }
+
+    loadAndClose() {
+        this.load().then(() => this.props.onTryClose());
+    }
+
+    loadAs() {
+        this.redefineVariableName()
+            .then(() => {
+                this.loadVariable();
+            })
+    }
+
+    variableNameExists() {
+        return Promise.resolve(this.variableName in this.props.curVariables);
+    }
+
+    redefineVariableName() {
+        return new Promise((resolve, reject) => {
+            this.doneRedefineVariable = () => { resolve() };
+            this.setState({
+                showRedefineVariableModal: true,
+                temporaryRedefinedVariableName: this.variableName
+            });
+        })
     }
 
     handleFileExplorerTryClose() {
@@ -109,6 +215,174 @@ class FileTab extends Component{
 
     handleDragEnd(){
         this.setState({showBookmarkZone: false})
+    }
+
+    render() {
+        return (
+            <div>
+                <Modal.Body>
+                    <TabBar switchTab={this.props.switchTab} selectedTab={this.props.selectedTab}/>
+                    <div className="load-from">
+                        <Row>
+                            <Col className="text-right" sm={2}>
+                                <h4>Load From</h4>
+                            </Col>
+                        </Row>
+                        <Row>
+                            <Col className="text-right" sm={2}>
+                                File
+                            </Col>
+                            <Col sm={9}>
+                                <InputGroup bsSize="small">
+                                    <FormControl className="full-width file-path" type="text" value={this.selectedFilePath} />
+                                    <InputGroup.Button>
+                                        <Button bsStyle="primary" onClick={
+                                            () => this.setState({ showFileExplorer: true })}><Glyphicon glyph="file" /></Button>
+                                    </InputGroup.Button>
+                                </InputGroup>
+                            </Col>
+                        </Row>
+                        <Row>
+                            <Col className="text-right" sm={2}>
+                                Variable(s)
+                            </Col>
+                            <Col sm={9}>
+                                <FormControl
+                                    className="input-sm full-width"
+                                    componentClass="select"
+                                    onChange={(e) => this.setState({ selectedVariableName: e.target.value })}
+                                    value={this.state.selectedVariableName}>
+                                    {this._formatvariablesAxes()}
+                                </FormControl>
+                            </Col>
+                        </Row>
+                        <Row>
+                            <Col className="text-right" sm={2}>
+                                History:
+                            </Col>
+                            <Col sm={9}>
+                                <FormControl className="history" componentClass="div">
+                                    {this.state.historyFiles.map((file, i) => {
+                                        return (
+                                            <div 
+                                                className="file"
+                                                key={i}
+                                                draggable="true"
+                                                onDragStart={(e) => {this.handleDragStart(e, file)}}
+                                                onDragEnd={(e) => {this.handleDragEnd(e)}}
+                                                onClick={(e) => this.handleFileSelected(file)}>
+                                                {cleanPath(file.path + '/' + file.name)}
+                                            </div>
+                                        )
+                                    })}
+                                </FormControl>
+                            </Col>
+                        </Row>
+                        <Row>
+                            <Col className="text-right" sm={2}>
+                                Bookmarks:
+                            </Col>
+                            <Col sm={9}>
+                                <FormControl 
+                                    className="bookmarks"
+                                    componentClass="div"
+                                    style={{backgroundColor: this.state.showBookmarkZone ? "#d1ecf1" : "#fff"}}
+                                    onDragOver={(e) => {this.handleDragOver(e)}}
+                                    onDrop={(e) => {this.handleDrop(e)}}
+                                    >
+                                    {this.state.bookmarkFiles.map((file, i) => {
+                                        return( 
+                                            <div
+                                                className="file" 
+                                                key={i} 
+                                                onClick={(e) => this.handleFileSelected(file)}>
+                                                {cleanPath(file.path + '/' + file.name)}
+                                            </div>
+                                        )
+                                    })}
+                                </FormControl>
+                            </Col>
+                        </Row>
+                    </div>
+                    {this.state.selectedVariable &&
+                        <div className="dimensions">
+                            <Row>
+                                <Col className="text-right" sm={2}>
+                                    <h4>Dimensions</h4>
+                                </Col>
+                            </Row>
+                            {/* If is a variable */}
+                            {this.state.dimension &&
+                                this.state.dimension.map(dimension => dimension.axisName).map((axisName, i) => {
+                                    let axis = this.state.variablesAxes[1][axisName];
+                                    return (
+                                        <DimensionDnDContainer key={axisName} index={i} axis={axis} axisName={axisName} handleDimensionValueChange={(values) => this.handleDimensionValueChange(values, axisName)} moveDimension={(dragIndex, hoverIndex) => this.moveDimension(dragIndex, hoverIndex)} />
+                                    )
+                                })
+                            }
+                            {/* if is an Axis */}
+                            {!this.state.selectedVariable.axisList &&
+                                <Row key={this.state.selectedVariable.name} className="dimension">
+                                    <Col sm={2} className="text-right"><span>{this.state.selectedVariable.name}</span></Col>
+                                    <Col sm={8} className="right-content">
+                                        <DimensionSlider {...this.state.selectedVariable} onChange={(values) => this.handleDimensionValueChange(values)} />
+                                    </Col>
+                                </Row>
+                            }
+                        </div>
+                    }
+                </Modal.Body>
+                <Modal.Footer>
+                    <Button bsStyle="primary" bsSize="small" onClick={(e) => {
+                        this.load()
+                    }}>Load</Button>
+                    <Button bsStyle="primary" bsSize="small" onClick={(e) => {
+                        this.loadAndClose()
+                    }}>Load and Close</Button>
+                    <Button bsStyle="primary" bsSize="small" onClick={(e) => {
+                        this.loadAs()
+                    }}>Load As</Button>
+                    <Button bsStyle="default" bsSize="small" onClick={() => this.props.onTryClose()}>Close</Button>
+                </Modal.Footer>
+                <Modal show={this.state.showRedefineVariableModal}>
+                    <Modal.Header>
+                        <Modal.Title id="contained-modal-title-sm">Rename variable</Modal.Title>
+                    </Modal.Header>
+                    <Modal.Body>
+                        <FormGroup>
+                            <ControlLabel>Variable name</ControlLabel>
+                            <FormControl
+                                type="text"
+                                value={this.state.temporaryRedefinedVariableName}
+                                onChange={(e) => this.setState({ temporaryRedefinedVariableName: e.target.value })} />
+                        </FormGroup>
+                    </Modal.Body>
+                    <Modal.Footer>
+                        <Button bsStyle="primary" bsSize="small"
+                            onClick={() => {
+                                if (this.state.temporaryRedefinedVariableName) {
+                                    this.doneRedefineVariable();
+                                    this.setState({
+                                        redefinedVariableName: this.state.temporaryRedefinedVariableName,
+                                        showRedefineVariableModal: false
+                                    });
+                                }
+                            }}>Confirm</Button>
+                        <Button bsSize="small" onClick={() => this.setState({ showRedefineVariableModal: false })}>Close</Button>
+                    </Modal.Footer>
+                </Modal>
+                <Dialog ref="dialog" />
+                {this.state.showFileExplorer &&
+                    <FileExplorer show={true} onTryClose={() => this.handleFileExplorerTryClose()} onFileSelected={(file) => this.handleFileSelected(file)} />}
+            </div>
+        )
+    }
+
+    _formatvariablesAxes() {
+        return this.state.variablesAxes && [
+            <optgroup key="variables" label="------">{this._formatVariables(this.state.variablesAxes[0])}</optgroup>,
+            <optgroup key="axes" label="------">{this._formatAxes(this.state.variablesAxes[1])}</optgroup>
+        ]
     }
 
     _formatVariables(variables) {
@@ -177,122 +451,24 @@ class FileTab extends Component{
         dimensions.splice(hoverIndex, 0, dimensions.splice(dragIndex, 1)[0]);
         this.setState({ dimension: dimensions });
     }
+}
+FileTab.propTypes = {
+    onTryClose: React.PropTypes.func.isRequired,
+    cachedFiles: React.PropTypes.object,
+    curVariables: React.PropTypes.object.isRequired,
+    loadVariables: React.PropTypes.func,
+    addFileToCache: React.PropTypes.func,
+}
 
-    render(){
-        return(
-            <div>
-                <div className="load-from">
-                    <Row>
-                        <Col className="text-right" sm={2}>
-                            <h4>Load From</h4>
-                        </Col>
-                    </Row>
-                    <Row>
-                        <Col className="text-right" sm={2}>
-                            File
-                        </Col>
-                        <Col sm={9}>
-                            <InputGroup bsSize="small">
-                                <FormControl className="full-width file-path" type="text" value={this.selectedFilePath} />
-                                <InputGroup.Button>
-                                    <Button bsStyle="primary" onClick={
-                                        () => this.setState({ showFileExplorer: true })}><Glyphicon glyph="file" /></Button>
-                                </InputGroup.Button>
-                            </InputGroup>
-                        </Col>
-                    </Row>
-                    <Row>
-                        <Col className="text-right" sm={2}>
-                            Variable(s)
-                        </Col>
-                        <Col sm={9}>
-                            <FormControl
-                                className="input-sm full-width"
-                                componentClass="select"
-                                onChange={(e) => this.setState({ selectedVariableName: e.target.value })}
-                                value={this.state.selectedVariableName}>
-                                {this._formatvariablesAxes()}
-                            </FormControl>
-                        </Col>
-                    </Row>
-                    <Row>
-                        <Col className="text-right" sm={2}>
-                            History:
-                        </Col>
-                        <Col sm={9}>
-                            <FormControl className="history" componentClass="div">
-                                {this.state.historyFiles.map((file, i) => {
-                                    return (
-                                        <div 
-                                            className="file"
-                                            key={i}
-                                            draggable="true"
-                                            onDragStart={(e) => {this.handleDragStart(e, file)}}
-                                            onDragEnd={(e) => {this.handleDragEnd(e)}}
-                                            onClick={(e) => this.handleFileSelected(file)}>
-                                            {cleanPath(file.path + '/' + file.name)}
-                                        </div>
-                                    )
-                                })}
-                            </FormControl>
-                        </Col>
-                    </Row>
-                    <Row>
-                        <Col className="text-right" sm={2}>
-                            Bookmarks:
-                        </Col>
-                        <Col sm={9}>
-                            <FormControl 
-                                className="bookmarks"
-                                componentClass="div"
-                                style={{backgroundColor: this.state.showBookmarkZone ? "#d1ecf1" : "#fff"}}
-                                onDragOver={(e) => {this.handleDragOver(e)}}
-                                onDrop={(e) => {this.handleDrop(e)}}
-                                >
-                                {this.state.bookmarkFiles.map((file, i) => {
-                                    return( 
-                                        <div
-                                            className="file" 
-                                            key={i} 
-                                            onClick={(e) => this.handleFileSelected(file)}>
-                                            {cleanPath(file.path + '/' + file.name)}
-                                        </div>
-                                    )
-                                })}
-                            </FormControl>
-                        </Col>
-                    </Row>
-                </div>
-                {this.state.selectedVariable &&
-                    <div className="dimensions">
-                        <Row>
-                            <Col className="text-right" sm={2}>
-                                <h4>Dimensions</h4>
-                            </Col>
-                        </Row>
-                        {/* If is a variable */}
-                        {this.state.dimension &&
-                            this.state.dimension.map(dimension => dimension.axisName).map((axisName, i) => {
-                                let axis = this.state.variablesAxes[1][axisName];
-                                return (
-                                    <DimensionDnDContainer key={axisName} index={i} axis={axis} axisName={axisName} handleDimensionValueChange={(values) => this.handleDimensionValueChange(values, axisName)} moveDimension={(dragIndex, hoverIndex) => this.moveDimension(dragIndex, hoverIndex)} />
-                                )
-                            })
-                        }
-                        {/* if is an Axis */}
-                        {!this.state.selectedVariable.axisList &&
-                            <Row key={this.state.selectedVariable.name} className="dimension">
-                                <Col sm={2} className="text-right"><span>{this.state.selectedVariable.name}</span></Col>
-                                <Col sm={8} className="right-content">
-                                    <DimensionSlider {...this.state.selectedVariable} onChange={(values) => this.handleDimensionValueChange(values)} />
-                                </Col>
-                            </Row>
-                        }
-                    </div>
-                }
-            </div>
-        )
-    }
+var DimensionContainer = (props) => {
+    const opacity = props.isDragging ? 0 : 1;
+    return props.connectDropTarget(props.connectDragPreview(<div className="row dimension" style={{ opacity }}>
+        <Col sm={2} className="text-right"><span>{props.axis.name}</span></Col>
+        {props.connectDragSource(<div className="sort col-sm-1"><Glyphicon glyph="menu-hamburger" /></div>)}
+        <div className="col-sm-7 right-content">
+            <DimensionSlider {...props.axis} onChange={props.handleDimensionValueChange} />
+        </div>
+    </div>));
 }
 
 var DimensionDnDContainer = _.flow(
@@ -345,17 +521,8 @@ var DimensionDnDContainer = _.flow(
                 isOver: monitor.isOver(),
             };
         }
-    ))(DimensionContainer);
+    ))
+    (DimensionContainer);
 
-var DimensionContainer = (props) => {
-    const opacity = props.isDragging ? 0 : 1;
-    return props.connectDropTarget(props.connectDragPreview(<div className="row dimension" style={{ opacity }}>
-        <Col sm={2} className="text-right"><span>{props.axis.name}</span></Col>
-        {props.connectDragSource(<div className="sort col-sm-1"><Glyphicon glyph="menu-hamburger" /></div>)}
-        <div className="col-sm-7 right-content">
-            <DimensionSlider {...props.axis} onChange={props.handleDimensionValueChange} />
-        </div>
-    </div>));
-}
 
-export default FileTab
+export default FileTab;
